feat(service): Throw appropriate exception in method to delete review by id

+++ src/main/java/com/israelopeters/rtireviews/service/ReviewServiceImpl.java
@@ -57,6 +57,11 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public void deleteReview(Long id) {
-        reviewRepository.deleteById(id);
+        Optional<Review> existingReview = reviewRepository.findById(id);
+        if (existingReview.isPresent()) {
+            reviewRepository.deleteById(id);
+        }
+        throw new ReviewNotFoundException(
+                String.format("No review found with ID: %d", id));
     }
 }
